requirements.txt
# Core dependencies
accelerate==1.0.1
certifi==2025.8.3
charset-normalizer==3.4.3
filelock==3.16.1
fsspec==2025.3.0
hf-xet==1.1.9
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
packaging==25.0
pillow==10.4.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
psutil==7.0.0
pypdf==5.9.0
PyMuPDF==1.24.9
PyYAML==6.0.2
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==3.2.1
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.2.2
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.13.2
urllib3==2.2.3

# LLM + API clients
openai==1.30.1
streamlit==1.39.0
flask==3.0.3

# NLP evaluation metrics
nltk==3.9.1
rouge-score==0.1.2
